chore: readapt to base PR
