--- packages/fuels-programs/src/execution_script.rs
@@ -1,146 +1,0 @@
-use std::{fmt::Debug, vec};
-
-use fuel_tx::{AssetId, FormatValidityChecks, Receipt, Script, ScriptExecutionResult, Transaction};
-use fuels_core::{offsets::call_script_data_offset, parameters::TxParameters};
-use fuels_signers::{provider::Provider, Signer, WalletUnlocked};
-use fuels_types::errors::{Error, Result};
-use fuels_types::param_types::ParamType;
-
-use crate::{
-    call_utils::{
-        build_script_data_from_contract_calls, calculate_required_asset_amounts, get_instructions,
-        get_single_call_instructions, get_transaction_inputs_outputs, CallOpcodeParamsOffset,
-    },
-    contract::ContractCall,
-};
-
-/// [`ExecutableFuelCall`] provides methods to create and call/simulate a transaction that carries
-/// out contract method calls or script calls
-#[derive(Debug)]
-pub struct ExecutableFuelCall {
-    pub tx: Script,
-}
-
-impl ExecutableFuelCall {
-    pub fn new(tx: Script) -> Self {
-        Self { tx }
-    }
-
-    /// Creates a [`ExecutableFuelCall`] from contract calls. The internal [Transaction] is
-    /// initialized with the actual script instructions, script data needed to perform the call and
-    /// transaction inputs/outputs consisting of assets and contracts.
-    pub async fn from_contract_calls(
-        calls: &[ContractCall],
-        tx_parameters: &TxParameters,
-        wallet: &WalletUnlocked,
-    ) -> Result<Self> {
-        let consensus_parameters = wallet.get_provider()?.consensus_parameters().await?;
-        let n_vectors_calls = calls.iter().filter(|c| c.output_param.is_vector()).count();
-
-        // Compute the length of the calling scripts for the two types of contract calls.
-        // Use placeholder for `call_param_offsets` and `output_param_type`, because the length of
-        // the script doesn't depend on the underlying data, just on whether or not the contract
-        // call output is a vector.
-        let calls_instructions_len_no_vectors =
-            get_single_call_instructions(&CallOpcodeParamsOffset::default(), &ParamType::U64).len()
-                * (calls.len() - n_vectors_calls);
-        let calls_instructions_len_vectors = get_single_call_instructions(
-            &CallOpcodeParamsOffset::default(),
-            &ParamType::Vector(Box::from(ParamType::U64)),
-        )
-        .len()
-            * n_vectors_calls;
-
-        let calls_instructions_len =
-            calls_instructions_len_no_vectors + calls_instructions_len_vectors;
-
-        let data_offset = call_script_data_offset(&consensus_parameters, calls_instructions_len);
-
-        let (script_data, call_param_offsets) =
-            build_script_data_from_contract_calls(calls, data_offset, tx_parameters.gas_limit);
-
-        let script = get_instructions(calls, call_param_offsets);
-
-        let required_asset_amounts = calculate_required_asset_amounts(calls);
-        let mut spendable_resources = vec![];
-
-        // Find the spendable resources required for those calls
-        for (asset_id, amount) in &required_asset_amounts {
-            let resources = wallet.get_spendable_resources(*asset_id, *amount).await?;
-            spendable_resources.extend(resources);
-        }
-
-        let (inputs, outputs) =
-            get_transaction_inputs_outputs(calls, wallet.address(), spendable_resources);
-
-        let mut tx = Transaction::script(
-            tx_parameters.gas_price,
-            tx_parameters.gas_limit,
-            tx_parameters.maturity,
-            script,
-            script_data,
-            inputs,
-            outputs,
-            vec![],
-        );
-
-        let base_asset_amount = required_asset_amounts
-            .iter()
-            .find(|(asset_id, _)| *asset_id == AssetId::default());
-        match base_asset_amount {
-            Some((_, base_amount)) => wallet.add_fee_resources(&mut tx, *base_amount, 0).await?,
-            None => wallet.add_fee_resources(&mut tx, 0, 0).await?,
-        }
-        wallet.sign_transaction(&mut tx).await.unwrap();
-
-        Ok(ExecutableFuelCall::new(tx))
-    }
-
-    /// Execute the transaction in a state-modifying manner.
-    pub async fn execute(&self, provider: &Provider) -> Result<Vec<Receipt>> {
-        let chain_info = provider.chain_info().await?;
-
-        self.tx.check_without_signatures(
-            chain_info.latest_block.header.height,
-            &chain_info.consensus_parameters,
-        )?;
-
-        provider.send_transaction(&self.tx).await
-    }
-
-    /// Execute the transaction in a simulated manner, not modifying blockchain state
-    pub async fn simulate(&self, provider: &Provider) -> Result<Vec<Receipt>> {
-        let chain_info = provider.chain_info().await?;
-
-        self.tx.check_without_signatures(
-            chain_info.latest_block.header.height,
-            &chain_info.consensus_parameters,
-        )?;
-
-        let receipts = provider.dry_run(&self.tx.clone().into()).await?;
-        Self::validate_script_succedded(&receipts)?;
-
-        Ok(receipts)
-    }
-
-    fn validate_script_succedded(receipts: &[Receipt]) -> Result<()> {
-        receipts
-            .iter()
-            .find_map(|receipt| match receipt {
-                Receipt::ScriptResult { result, .. }
-                    if *result != ScriptExecutionResult::Success =>
-                {
-                    Some(format!("{result:?}"))
-                }
-                _ => None,
-            })
-            .map(|error_message| {
-                Err(Error::RevertTransactionError {
-                    reason: error_message,
-                    revert_id: 0,
-                    receipts: receipts.to_owned(),
-                })
-            })
-            .unwrap_or(Ok(()))
-    }
-}
