+++ packages/fuels-signers/src/provider.rs
@@ -91,12 +91,13 @@ impl Provider {
     /// ## Sending a transaction
     ///
     /// ```
+    /// use std::default::Default;
     /// use fuels::tx::Script;
     /// use fuels::prelude::*;
     /// async fn foo() -> std::result::Result<(), Box<dyn std::error::Error>> {
     ///   // Setup local test node
     ///   let (provider, _) = setup_test_provider(vec![], vec![], None, None).await;
-    ///   let tx = Script::default();
+    ///   let tx = ScriptTransaction::new(&[],&[], TxParameters::default());
     ///
     ///   let receipts = provider.send_transaction(&tx).await?;
     ///   dbg!(receipts);
